+++ src/renderer/bootstrap.tsx
@@ -31,22 +31,17 @@ import * as LensExtensionsRendererApi from "../extensions/renderer-api";
 import { render, unmountComponentAtNode } from "react-dom";
 import { delay } from "../common/utils";
 import { isMac, isDevelopment } from "../common/vars";
-import { HotbarStore } from "../common/hotbar-store";
 import { ClusterStore } from "../common/cluster-store";
 import { UserStore } from "../common/user-store";
 import { ExtensionDiscovery } from "../extensions/extension-discovery";
 import { ExtensionLoader } from "../extensions/extension-loader";
-import { ExtensionsStore } from "../extensions/extensions-store";
-import { FilesystemProvisionerStore } from "../main/extension-filesystem";
 import { App } from "./components/app";
 import { LensApp } from "./lens-app";
-import { ThemeStore } from "./theme.store";
 import { HelmRepoManager } from "../main/helm/helm-repo-manager";
 import { ExtensionInstallationStateStore } from "./components/+extensions/extension-install.store";
 import { DefaultProps } from "./mui-base-theme";
 import configurePackages from "../common/configure-packages";
 import * as initializers from "./initializers";
-import { WeblinkStore } from "../common/weblink-store";
 
 configurePackages();
 
@@ -83,30 +78,13 @@ export async function bootstrap(App: AppComponent) {
   ExtensionLoader.createInstance().init();
   ExtensionDiscovery.createInstance().init();
 
-  const userStore = UserStore.createInstance();
-  const clusterStore = ClusterStore.createInstance();
-  const extensionsStore = ExtensionsStore.createInstance();
-  const filesystemStore = FilesystemProvisionerStore.createInstance();
-  const themeStore = ThemeStore.createInstance();
-  const hotbarStore = HotbarStore.createInstance();
-  const weblinkStore = WeblinkStore.createInstance();
+  await initializers.initStores();
 
   ExtensionInstallationStateStore.bindIpcListeners();
   HelmRepoManager.createInstance(); // initialize the manager
 
-  // preload common stores
-  await Promise.all([
-    userStore.load(),
-    hotbarStore.load(),
-    clusterStore.load(),
-    extensionsStore.load(),
-    filesystemStore.load(),
-    themeStore.init(),
-    weblinkStore.load()
-  ]);
-
   // Register additional store listeners
-  clusterStore.registerIpcListener();
+  ClusterStore.getInstance().registerIpcListener();
 
   // init app's dependencies if any
   if (App.init) {
